requirements.txt
numpy~=2.2.5
pillow~=11.2.1
argparse~=1.4.0